+++ frontend/src/pages/Overview/components/UserSummary/style.ts
@@ -79,21 +79,3 @@ export const FavoriteChampionsWrapper = styled(Box)(({ theme }) => ({
         }
     }
 }));
-
-export const CircleChartWrapper = styled('div')(({ theme }) => ({
-    position: "relative",
-    width: "fit-content"
-}));
-
-// Do we want to create our own title using div, similar to labels?
-
-export const LabelWrapper = styled('div')(({ theme }) => ({
-    position: "absolute",
-    bottom: 0,
-    display: "flex",
-    flexDirection: "column",
-    alignItems: "center",
-    justifyContent: "center",
-    height: '100%',
-    width: "100%",
-}));
