+++ frontend/src/redux/slices/user.ts
@@ -2,22 +2,25 @@ import { createSlice, PayloadAction, Dispatch } from "@reduxjs/toolkit";
 import { UserState } from '../../interfaces'
 import axios from 'axios';
 
-const initialState: UserState = {}
+const initialState: UserState = {
+    summonerName: 'Player',
+    summonerLevel: 0,
+    summonerIcon:  undefined,
+    rank:          'IRON',
+    leaguePoints:  0,
+    winGames:      0,
+    lossGames:     0,
+    favChamps:     [],
+    region:        'NA',
+    summonerFound: false
+}
 
 const userSlice = createSlice({
     name: 'user',
     initialState,
     reducers: {
         setUserData: (state, action: PayloadAction<UserState>) => {
-            state.summonerName  = action.payload.summonerName   ?? undefined;
-            state.summonerLevel = action.payload.summonerLevel  ?? undefined;
-            state.summonerIcon  = action.payload.summonerIcon   ?? undefined;
-            state.rank          = action.payload.rank           ?? undefined;
-            state.winGames      = action.payload.winGames       ?? undefined;
-            state.lossGames     = action.payload.lossGames      ?? undefined;
-            state.favChamps     = action.payload.favChamps      ?? undefined;
-            state.region        = action.payload.region         ?? undefined;
-            state.summonerFound = action.payload.summonerFound  ?? undefined;
+            return { ...action.payload };
         },
     },
 })
@@ -27,7 +30,7 @@ function fetchUserData(query: string) {
         if (query) {
             const params = new URLSearchParams(query);
             const summonerName = params.get('summonerName');
-            const region = params.get('region') ?? "NA"; // Defaults to NA region
+            const region = params.get('region') ?? 'NA'; // Defaults to NA region
             console.log(region);
             if (summonerName) {
                 axios.get('/api/summonerInfo', {
@@ -39,11 +42,11 @@ function fetchUserData(query: string) {
                 .then((res) => {
                     const userData = res.data;
                     console.log('user found:', userData);
-                    dispatch(setUserData({summonerFound: true, region: region, ...userData}));
+                    dispatch(setUserData({ summonerFound: true, region: region, ...userData }));
                 })
                 .catch((err) => {
                     console.log('user not found:\n', err.response || err);
-                    dispatch(setUserData({summonerFound: false, region: region, summonerName: summonerName}));
+                    dispatch(setUserData({ summonerFound: false, region: region, ...initialState }));
                 });
             }
         }
